third_party/WebKit/LayoutTests/editing/spelling/markers-expected.txt

Grammar Error Only
PASS internals.markerCountForNode(e.firstChild, "spelling") is 0
FAIL internals.markerCountForNode(e.firstChild, "grammar") should be 1. Was 0.
FAIL range.toString() should be a. Threw exception TypeError: Cannot call method 'toString' of null
Spelling Error Only
PASS internals.markerCountForNode(e.firstChild, "spelling") is 1
PASS range.toString() is "zz"
PASS internals.markerCountForNode(e.firstChild, "grammar") is 0
Grammar and Spelling Error
PASS internals.markerCountForNode(e.firstChild, "spelling") is 1
PASS range.toString() is "zz"
FAIL internals.markerCountForNode(e.firstChild, "grammar") should be 1. Was 0.
FAIL range.toString() should be orange,zz,apple.. Threw exception TypeError: Cannot call method 'toString' of null
PASS successfullyParsed is true

TEST COMPLETE
I have a issue.
zz.
orange,zz,apple.
